services/worker/pyproject.toml
[project]
name = "dijkfood-worker"
version = "0.1.0"
dependencies = [
    "fastapi[standard]",
    "boto3",
    "networkx",
    "numpy",
    "shapely",
    "psycopg2-binary",
]

[tool.uv]
managed = true